Feature/rotate weapon (#109)
* added rmb

* added priority to rendering the animations

also added rednerpriority to weapons to be higher

* save commit 1

* weapon showing v1

* aiming fix

* aiming working

* aiming fixed, weapon positition fixed

it is all done

* clang fixes

* refactored input to double

* Fix crash

* refactor fixes

* Rotate weapons around a point

* more fixes

* morefixes

* clang fixes

* feedback and clang

* Add renderpriority

* fixed magic value

* clang and added controller removal support

that support does not seem to work correctly

* Remove commented out code

* Use joystick intensity to calculate movement speed

* clang fix and optimized

* fixed bug in inventorysysyem and fixed knockback

* clang

* added some comments

* tried fixing the fps

* Clear inventory on death

* fixed all

* fixed stuff

* Add rendering priority to props

* clang fixes

+++ engine/events/models/ControllerEvent.h
@@ -10,7 +10,7 @@ namespace events {
         int m_ID;
         input::Keys m_key;
         input::AnalogKeys m_analogKey;
-        int m_axisValue;
+        double m_axisValue;
         bool m_keyDown;
         bool m_isAnalog = false;
 
@@ -19,7 +19,7 @@ namespace events {
             , m_key{ key }
             , m_keyDown{ keyDown } {};
 
-        ControllerEvent(int ID, input::AnalogKeys analogKey, int axisValue)
+        ControllerEvent(int ID, input::AnalogKeys analogKey, double axisValue)
             : m_ID(ID)
             , m_analogKey(analogKey)
             , m_axisValue(axisValue)
